refactor: move register flow schemas to auth.ts in shared schemas
