+++ packages/ui/src/components/Generic/RegisterFlow/BasicInfoForm2/BasicInfoForm2.tsx
@@ -1,11 +1,11 @@
 "use client"
 
 import { zodResolver } from "@hookform/resolvers/zod"
+import { BasicInfoForm2Schema, type BasicInfoForm2Values } from "@repo/shared"
 import { Button, Heading, InputType, TextInput } from "@repo/ui/components/Primitive"
 import { FormControl, memo, noop, VStack } from "@yamada-ui/react"
 import type { FC } from "react"
 import { type SubmitHandler, useForm } from "react-hook-form"
-import { BasicInfoForm2Schema, type BasicInfoForm2Values } from "../schema"
 
 /**
  * Props for {@link BasicInfoForm2} component
